Change the gamut mask docker to use KisAngleSelector

+++ libs/ui/widgets/KisGamutMaskToolbar.h
@@ -47,7 +47,7 @@ public Q_SLOTS:
 
 private Q_SLOTS:
     void slotGamutMaskToggle(bool state);
-    void slotGamutMaskRotate(int angle);
+    void slotGamutMaskRotate(qreal angle);
 
 private:
     QScopedPointer<Ui_wdgGamutMaskToolbar> m_ui;
